# src/api/swagger.json
{
	"basePath": "/v1",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ResponseModel": {
			"properties": {
				"status": {
					"type": "number",
					"format": "double"
				},
				"message": {
					"type": "string"
				}
			},
			"required": [
				"status",
				"message"
			],
			"type": "object"
		},
		"BroadlinkDeviceInfo": {
			"properties": {
				"id": {
					"type": "number",
					"format": "double"
				},
				"host": {
					"type": "string"
				},
				"mac": {
					"type": "string"
				},
				"type": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"host",
				"type"
			],
			"type": "object"
		},
		"BroadlinkRequest": {
			"properties": {
				"device": {
					"type": "string",
					"enum": [
						"Projector",
						"TV"
					]
				},
				"command": {
					"type": "string"
				}
			},
			"required": [
				"device",
				"command"
			],
			"type": "object"
		},
		"ChromecastOperations": {
			"enum": [
				"play",
				"pause",
				"resume",
				"stop",
				"seek",
				"subtitles"
			],
			"type": "string"
		},
		"ChromecastMediaOptions": {
			"properties": {
				"title": {
					"type": "string"
				},
				"type": {
					"type": "string"
				},
				"seek": {
					"type": "number",
					"format": "double"
				},
				"subtitles": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"autoSubtitles": {
					"type": "boolean"
				}
			},
			"type": "object"
		},
		"ChromecastRequest": {
			"properties": {
				"command": {
					"$ref": "#/definitions/ChromecastOperations"
				},
				"url": {
					"type": "string"
				},
				"options": {
					"$ref": "#/definitions/ChromecastMediaOptions"
				}
			},
			"required": [
				"command"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "tars",
		"version": "1.0.0",
		"description": "The Automated Residence System",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/Broadlink/Discover": {
			"post": {
				"operationId": "DiscoverBroadlinkDevices",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ResponseModel"
						}
					}
				},
				"tags": [
					"Broadlink"
				],
				"security": [],
				"parameters": []
			}
		},
		"/Broadlink/Devices": {
			"get": {
				"operationId": "GetBroadlinkDevices",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					}
				},
				"tags": [
					"Broadlink"
				],
				"security": [],
				"parameters": []
			}
		},
		"/Broadlink/Devices/{deviceId}": {
			"get": {
				"operationId": "GetBroadlinkDevice",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/BroadlinkDeviceInfo"
						}
					}
				},
				"tags": [
					"Broadlink"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/Broadlink/Devices/{deviceId}/Controller": {
			"put": {
				"operationId": "PutBroadlinkDeviceController",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "No content",
						"schema": {
							"$ref": "#/definitions/ResponseModel"
						}
					}
				},
				"tags": [
					"Broadlink"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "body",
						"name": "model",
						"required": true,
						"schema": {
							"$ref": "#/definitions/BroadlinkRequest"
						}
					}
				]
			}
		},
		"/HealthCheck": {
			"get": {
				"operationId": "GetHealth",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ResponseModel"
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/Chromecast/Discover": {
			"post": {
				"operationId": "DiscoverChromecasts",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					}
				},
				"tags": [
					"Chromecast"
				],
				"security": [],
				"parameters": []
			}
		},
		"/Chromecast/Devices": {
			"get": {
				"operationId": "GetChromecastDevices",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					}
				},
				"tags": [
					"Chromecast"
				],
				"security": [],
				"parameters": []
			}
		},
		"/Chromecast/Devices/{deviceId}": {
			"get": {
				"operationId": "GetChromecastDevice",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					}
				},
				"tags": [
					"Chromecast"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/Chromecast/Devices/{deviceId}/Controller": {
			"put": {
				"operationId": "PutChromecastDeviceController",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "No content",
						"schema": {
							"$ref": "#/definitions/ResponseModel"
						}
					}
				},
				"tags": [
					"Chromecast"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "model",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ChromecastRequest"
						}
					}
				]
			}
		},
		"/Chromecast/Default/Controller": {
			"put": {
				"operationId": "ControlChromecast",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ResponseModel"
						}
					}
				},
				"tags": [
					"Chromecast"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ChromecastRequest"
						}
					}
				]
			}
		},
		"/Chromecast/Default": {
			"put": {
				"operationId": "SetDefaultChromecastDevice",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ResponseModel"
						}
					}
				},
				"tags": [
					"Chromecast"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "device_name",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/Chromecast/Status": {
			"get": {
				"operationId": "GetChromecastDeviceStatus",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ResponseModel"
						}
					}
				},
				"tags": [
					"Chromecast"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "device_name",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
	"host": "localhost:3000"
}